Change Integer to Long

+++ apps/move-rows-to-shards/src/main/java/org/mediacloud/mrts/MinMaxTruncateActivities.java
@@ -10,11 +10,11 @@ public interface MinMaxTruncateActivities {
 
     @ActivityMethod
     @Nullable
-    Integer minColumnValue(String table, String idColumn);
+    Long minColumnValue(String table, String idColumn);
 
     @ActivityMethod
     @Nullable
-    Integer maxColumnValue(String table, String idColumn);
+    Long maxColumnValue(String table, String idColumn);
 
     @ActivityMethod
     void truncateIfEmpty(String table);
